Extract shared useLogBuffer hook from CrawlView/GenerateView
